refactor(tasks): migrate postprocess module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
