+++ src/pyrovelocity/tasks/__init__.py
@@ -1,10 +1,12 @@
 import pyrovelocity.tasks.data
+import pyrovelocity.tasks.postprocess
 import pyrovelocity.tasks.preprocess
 import pyrovelocity.tasks.train
 
 
 __all__ = [
     "data",
+    "postprocess",
     "preprocess",
     "train",
 ]
