+++ src/pyrovelocity/workflows/main_workflow.py
@@ -20,9 +20,9 @@ from pyrovelocity.interfaces import PyroVelocityTrainInterface
 from pyrovelocity.io.archive import create_tarball_from_filtered_dir
 from pyrovelocity.io.gcs import upload_file_concurrently
 from pyrovelocity.logging import configure_logging
-from pyrovelocity.postprocess import postprocess_dataset
 from pyrovelocity.summarize import summarize_dataset
 from pyrovelocity.tasks.data import download_dataset
+from pyrovelocity.tasks.postprocess import postprocess_dataset
 from pyrovelocity.tasks.preprocess import preprocess_dataset
 from pyrovelocity.tasks.train import train_dataset
 from pyrovelocity.utils import str_to_bool
